Finalized pruning command

+++ modules/service/dataset-pruning-service.js
@@ -1,5 +1,8 @@
 const Models = require('../../models/index');
 const { QueryTypes } = require('sequelize');
+const sequelizeConfig = require('../../config/sequelizeConfig');
+
+const defaultBackupFolderPath = '/ot-node/backup/';
 
 class DatasetPruningService {
     /**
@@ -9,6 +12,8 @@ class DatasetPruningService {
     constructor(ctx) {
         this.logger = ctx.logger;
         this.graphStorage = ctx.graphStorage;
+        this.diskService = ctx.diskService;
+        this.config = ctx.config;
     }
 
     getIdsForPruning(
@@ -110,11 +115,13 @@ class DatasetPruningService {
     }
 
     async removeEntriesWithId(table, idArray) {
-        await Models[table].destroy({
-            where: {
-                id: { [Models.sequelize.Op.in]: idArray },
-            },
-        });
+        if (idArray && idArray.length > 0) {
+            await Models[table].destroy({
+                where: {
+                    id: { [Models.sequelize.Op.in]: idArray },
+                },
+            });
+        }
     }
 
     async removeDatasetsFromGraphDb(datasets) {
@@ -208,6 +215,121 @@ class DatasetPruningService {
             }
         });
         return repackedDatasets;
+    }
+
+    shouldPruneLowEstimatedValueDatasets() {
+        if (!this.config.dataset_pruning.low_estimated_value_datasets.enabled) {
+            return false;
+        }
+        if (this.diskService.folderExisists(defaultBackupFolderPath)
+            && this.diskService.getFolderSize(defaultBackupFolderPath) > 0) {
+            this.logger.warn('Detected ot-node backup on machine. Unable to prune low estimated value datasets!');
+            return;
+        }
+
+        const diskSpace = this.diskService.getRootFolderInfo();
+
+        const freeSpacePercentage = (100 * diskSpace.free) / diskSpace.total;
+
+        return freeSpacePercentage < this.config.dataset_pruning
+            .low_estimated_value_datasets.minimum_free_space_percentage;
+        // const arangoDbEngineFolderSize = this.diskService
+        //     .getFolderSize(this.config.database.engine_folder_path);
+        // const operationDbSize = this.diskService.getFileSize(sequelizeConfig.storage);
+        //
+        // const minimumArangoDbFolderSizeForPruning = 0.2 * diskSpace.total;
+        // const minimumRequiredFreeSpace = (arangoDbEngineFolderSize + operationDbSize) * 1.1;
+        //
+        // return arangoDbEngineFolderSize > minimumArangoDbFolderSizeForPruning
+        //     && diskSpace.free < minimumRequiredFreeSpace;
+    }
+
+    async findLowEstimatedValueDatasets() {
+        const queryString = 'select di.id as data_info_id, di.data_set_id, bid.status as bid_status' +
+            'bid.id as bid_id, (di.import_timestamp + bid.holding_time_in_minutes*60000) as expiry,' +
+            'offer.id as offer_id, pd.id as purchase_id,' +
+            'from data_info as di ' +
+            'left join offers as offer on di.data_set_id = offer.data_set_id ' +
+            'inner join bids as bid on di.data_set_id = bid.data_set_id ' +
+            'left join purchased_data as pd on di.data_set_id = pd.data_set_id' +
+            'order by (di.import_timestamp + bid.holding_time_in_minutes*60000)';
+        return Models.sequelize.query(queryString, { type: QueryTypes.SELECT });
+    }
+
+    /**
+     * Returns repacked dataset information in format:
+     * {
+     *     [data_set_id]: {
+     *         dataInfos: [],
+     *         bids: [],
+     *     }
+     * }
+     * @param datasets
+     * @returns {{}}
+     */
+    repackLowEstimatedValueDatasets(datasets) {
+        const repackedDatasets = {};
+        datasets.forEach((dataset) => {
+            if (!repackedDatasets[dataset.data_set_id]) {
+                repackedDatasets[dataset.data_set_id] = {
+                    dataInfos: [],
+                    bids: [],
+                };
+            }
+            // to speed up if already found offer,
+            // purchase or choosen bid just continue to next element
+            if (!repackedDatasets[dataset.data_set_id].expiry
+                || repackedDatasets[dataset.data_set_id].expiry < dataset.expiry) {
+                repackedDatasets[dataset.data_set_id].expiry = dataset.expiry;
+            }
+            const foundDataInfoId = repackedDatasets[dataset.data_set_id].dataInfos
+                .includes(dataset.data_info_id);
+            if (!foundDataInfoId) {
+                repackedDatasets[dataset.data_set_id].dataInfos.push(dataset.data_info_id);
+            }
+
+            if (dataset.bid_id) {
+                const foundBidId = repackedDatasets[dataset.data_set_id].bids
+                    .includes(dataset.bid_id);
+                if (!foundBidId) {
+                    repackedDatasets[dataset.data_set_id].bids.push(dataset.bid_id);
+                    if (dataset.bids.bid_status === 'CHOOSEN') {
+                        repackedDatasets[dataset.data_set_id].choosen = true;
+                    }
+                }
+            }
+            if (dataset.offer_id) {
+                repackedDatasets[dataset.data_set_id].foundOffer = true;
+            }
+            if (dataset.purchase_id) {
+                repackedDatasets[dataset.data_set_id].foundPurchase = true;
+            }
+        });
+        // sort by expiry date low to high
+        return repackedDatasets.sort((a, b) => a.expiry - b.expiry);
+    }
+
+    getLowEstimatedValueIdsForPruning(repackedDatasets) {
+        const idsForPruning = {
+            datasetsToBeDeleted: [],
+            dataInfoIdToBeDeleted: [],
+            bidIdToBeDeleted: [],
+        };
+
+        Object.keys(repackedDatasets).forEach((key) => {
+            const dataset = repackedDatasets[key];
+            if (!dataset.foundOffer && !dataset.foundPurchase && !dataset.choosen) {
+                idsForPruning.datasetsToBeDeleted.push({
+                    datasetId: key,
+                });
+                idsForPruning.dataInfoIdToBeDeleted.concat(dataset.dataInfoIds);
+                idsForPruning.bidIdToBeDeleted.concat(dataset.bidsIds);
+            }
+        });
+
+        return repackedDatasets
+            .splice(this.config.dataset_pruning
+                .low_estimated_value_datasets.batch_size_for_pruning);
     }
 }
